core/requirements.txt
aioredis==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
Django==3.2.16
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.11.1
django-js-asset==2.1.0
django-redis==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-api-response==0.0.30
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
h11==0.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
mutagen==1.46.0
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==23.3.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.20.0
whitenoise==6.2.0
zope.interface==6.1
